Passage par fichier pour encodage et décodage

+++ guillerme/05_cryptographie/TP/02_gmp/menu.h
@@ -71,6 +71,6 @@ FILE *getFile(const char *mod);
  *  Description:  
  * =====================================================================================
  */
-char *readFile(FILE *f, size_t *s);
+char *readFile(FILE *f, int *s);
 
 #endif
